NEW: added volunteers filter on emergency area. Location entity was deleted, now is a point in volunteer and emergency. WebMercatorSingleton is on his own folder

+++ src/main/java/com/tallerbd/backend/emergency/EmergencyDTO.java
@@ -20,9 +20,9 @@ public class EmergencyDTO{
 
     private String inCharge;
 
-    private Float longitude;
+    private Double longitude;
 
-    private Float latitude;
+    private Double latitude;
 
     // private String creatorId;
